Split TokenCell into container and component files
