+++ ui/app/components/app/token-cell/index.js
@@ -1,0 +1,1 @@
+export { default } from './token-cell.container'
